app/requirements.txt
Flask
mysql-connector
simplejson
rq
redis
requests